refactor(web): extract runtime helpers from types.ts to response-utils.ts
jsonResponse(), errorResponse(), and readBody() were runtime functions
mixed into a *types.ts file, violating the convention that types files
are pure compile-time types. Extract them to server/response-utils.ts
and update all 22 callers (server routes + routes-rex subdir + test).
