+++ packages/web/src/server/routes-cli-timeout.ts
@@ -13,7 +13,7 @@ import type { IncomingMessage, ServerResponse } from "node:http";
 import { readFileSync, writeFileSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import type { ServerContext } from "./types.js";
-import { jsonResponse, errorResponse, readBody } from "./types.js";
+import { jsonResponse, errorResponse, readBody } from "./response-utils.js";
 
 // ---------------------------------------------------------------------------
 // Types
