+++ packages/web/src/server/routes-rex/health.ts
@@ -4,7 +4,7 @@
 
 import type { IncomingMessage, ServerResponse } from "node:http";
 import type { ServerContext } from "../types.js";
-import { jsonResponse, errorResponse, readBody } from "../types.js";
+import { jsonResponse, errorResponse, readBody } from "../response-utils.js";
 import type { WebSocketBroadcaster } from "../websocket.js";
 import { loadPRD, savePRD } from "./shared.js";
