+++ packages/web/src/server/routes-rex/items.ts
@@ -7,7 +7,7 @@ import { appendFileSync } from "node:fs";
 import { join } from "node:path";
 import { randomUUID } from "node:crypto";
 import type { ServerContext } from "../types.js";
-import { jsonResponse, errorResponse, readBody } from "../types.js";
+import { jsonResponse, errorResponse, readBody } from "../response-utils.js";
 import type { WebSocketBroadcaster } from "../websocket.js";
 import {
   findItemById, insertChild, updateInTree, loadPRD, savePRD,
