+++ packages/web/src/server/routes-rex/reads.ts
@@ -8,7 +8,7 @@ import type { ServerResponse } from "node:http";
 import { readFileSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import type { ServerContext } from "../types.js";
-import { jsonResponse, errorResponse } from "../types.js";
+import { jsonResponse, errorResponse } from "../response-utils.js";
 import { loadPRD, findNextTask, collectCompletedIds } from "./shared.js";
 
 import {
