+++ packages/web/src/server/routes-search.ts
@@ -15,7 +15,7 @@
 
 import type { IncomingMessage, ServerResponse } from "node:http";
 import type { ServerContext } from "./types.js";
-import { jsonResponse, errorResponse } from "./types.js";
+import { jsonResponse, errorResponse } from "./response-utils.js";
 import { SearchIndex } from "./search-index.js";
 
 // ── Index cache ────────────────────────────────────────────────────────────
