+++ packages/web/src/server/routes-status.ts
@@ -11,7 +11,7 @@ import type { IncomingMessage, ServerResponse } from "node:http";
 import { readFileSync, existsSync, readdirSync } from "node:fs";
 import { join } from "node:path";
 import type { ServerContext } from "./types.js";
-import { jsonResponse } from "./types.js";
+import { jsonResponse } from "./response-utils.js";
 import { DATA_FILES } from "../shared/index.js";
 import { computeStats, collectCompletedIds, findNextTask } from "./rex-gateway.js";
 import type { PRDDocument, TreeStats } from "./rex-gateway.js";
